Update beta branch purpose
Update of beta branch purpose. 
As breaking changes are incorporated into Production in v2022.0, these breaking changes should be available in Staging on 15/09/2021, so they can be tested and propagated to Production in the v2022.0 release. 
This makes the beta branch not needed for the period 15/09/2021-15/01/2022, as breaking and non-breaking changes will be available in Staging since 15/09/2021.